requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Werkzeug==2.3.7
Pillow==10.0.0
pytesseract==0.3.10
SpeechRecognition==3.10.0
mysqlclient==2.2.0 